+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [0.3.1]
+### Added
+- `Query::fetch_one()`.
+
 ### Changed
 - `Error` is `StdError + Send + Sync + 'static` now.